Rename MapObstacleComponent to MapObstacleEntity
