+++ core/src/main/java/io/github/alltheeb5t/unisim/factories/ObstaclesFactory.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Image;
 
 import io.github.alltheeb5t.unisim.building_components.BoundingBoxComponent;
 import io.github.alltheeb5t.unisim.entities.LibGdxRenderingEntity;
-import io.github.alltheeb5t.unisim.map_objects.MapObstacleComponent;
+import io.github.alltheeb5t.unisim.entities.MapObstacleEntity;
 
 public class ObstaclesFactory {
 
@@ -20,9 +20,9 @@ public class ObstaclesFactory {
      * @param campusMap
      * @return
      */
-    public static List<MapObstacleComponent> makeMapOrchard(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
+    public static List<MapObstacleEntity> makeMapOrchard(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
 
-        List<MapObstacleComponent> orchard = new LinkedList<>();
+        List<MapObstacleEntity> orchard = new LinkedList<>();
         int numOfTrees = 17;
 
         for (int i = 0; i < numOfTrees; i++) {
@@ -33,12 +33,12 @@ public class ObstaclesFactory {
             Image nextTreeTopImage = ImageComponentFactory.makeImageComponent(positionX, positionY, 5, new Texture("assets/MapObstacles/tree_square.png"));
             BoundingBoxComponent nextTreeTopBoundingBox = BoundingBoxComponentFactory.makeBoundingBoxComponent(nextTreeTopImage);
             libGdxRenderingEntity.getStage().addActor(nextTreeTopImage);
-            orchard.add(new MapObstacleComponent(nextTreeTopBoundingBox, nextTreeTopImage));
+            orchard.add(new MapObstacleEntity(nextTreeTopBoundingBox, nextTreeTopImage));
 
             Image nextTreeTrunkImage = ImageComponentFactory.makeImageComponent(positionX, positionY -2.5f, 1, 2, new Texture("assets/MapObstacles/tree_trunk_square.png"));
             BoundingBoxComponent nextTreeTrunkBoundingBox = BoundingBoxComponentFactory.makeBoundingBoxComponent(nextTreeTrunkImage);
             libGdxRenderingEntity.getStage().addActor(nextTreeTrunkImage);
-            orchard.add(new MapObstacleComponent(nextTreeTrunkBoundingBox, nextTreeTrunkImage));
+            orchard.add(new MapObstacleEntity(nextTreeTrunkBoundingBox, nextTreeTrunkImage));
         }
 
         return orchard;
@@ -51,12 +51,12 @@ public class ObstaclesFactory {
      * @param campusMap
      * @return
      */
-    public static MapObstacleComponent makeMapRoad(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) { 
+    public static MapObstacleEntity makeMapRoad(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) { 
         
         Image roadImage = ImageComponentFactory.makeImageComponent(x, y, 15, 2000, new Texture("assets/MapObstacles/road_square.png"));
         BoundingBoxComponent roadBoundingBoxComponent = BoundingBoxComponentFactory.makeBoundingBoxComponent(roadImage);
         libGdxRenderingEntity.getStage().addActor(roadImage);
-        return new MapObstacleComponent(roadBoundingBoxComponent, roadImage);
+        return new MapObstacleEntity(roadBoundingBoxComponent, roadImage);
     }
 
     /**
@@ -67,9 +67,9 @@ public class ObstaclesFactory {
      * @param round     determines the shape of the lake - larger number is more skewed lake
      * @return
      */
-    public static List<MapObstacleComponent> makeMapLake(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity, int round) {
+    public static List<MapObstacleEntity> makeMapLake(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity, int round) {
 
-        List<MapObstacleComponent> lake = new LinkedList<>();
+        List<MapObstacleEntity> lake = new LinkedList<>();
         
         int rowNumSquares = 3;
         float newX = x;
@@ -80,7 +80,7 @@ public class ObstaclesFactory {
                 Image nextLakeImage = ImageComponentFactory.makeImageComponent(newX, newY, 10, 5, new Texture("assets/MapObstacles/lake_square.png"));
                 BoundingBoxComponent nextLakeBoundingBoxComponent = BoundingBoxComponentFactory.makeBoundingBoxComponent(nextLakeImage);
                 libGdxRenderingEntity.getStage().addActor(nextLakeImage);
-                lake.add(new MapObstacleComponent(nextLakeBoundingBoxComponent, nextLakeImage));
+                lake.add(new MapObstacleEntity(nextLakeBoundingBoxComponent, nextLakeImage));
             }
 
             newY -= 5;
@@ -108,9 +108,9 @@ public class ObstaclesFactory {
      * @param campusMap
      * @return
      */
-    public static List<MapObstacleComponent> makeMapMountain(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
+    public static List<MapObstacleEntity> makeMapMountain(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
 
-        List<MapObstacleComponent> mountain = new LinkedList<>();
+        List<MapObstacleEntity> mountain = new LinkedList<>();
         
         int width = 0;
         float newY = y + 5;
@@ -130,7 +130,7 @@ public class ObstaclesFactory {
 
             BoundingBoxComponent nextLevelBoundingBoxComponent = BoundingBoxComponentFactory.makeBoundingBoxComponent(nextLevelImage);
 
-            mountain.add(new MapObstacleComponent(nextLevelBoundingBoxComponent, nextLevelImage));
+            mountain.add(new MapObstacleEntity(nextLevelBoundingBoxComponent, nextLevelImage));
             libGdxRenderingEntity.getStage().addActor(nextLevelImage);
         }
 
@@ -144,12 +144,12 @@ public class ObstaclesFactory {
      * @param campusMap
      * @return
      */
-    public static MapObstacleComponent makeMapRiver(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
+    public static MapObstacleEntity makeMapRiver(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
 
         Image riverImage = ImageComponentFactory.makeImageComponent(x, y, 10, 800, new Texture("assets/MapObstacles/lake_square.png"));
         BoundingBoxComponent riverBoundingBoxComponent = BoundingBoxComponentFactory.makeBoundingBoxComponent(riverImage);
         libGdxRenderingEntity.getStage().addActor(riverImage);
-        return new MapObstacleComponent(riverBoundingBoxComponent, riverImage);
+        return new MapObstacleEntity(riverBoundingBoxComponent, riverImage);
     }
 
     /**
@@ -159,10 +159,10 @@ public class ObstaclesFactory {
      * @param campusMap
      * @return
      */
-    public static MapObstacleComponent makeMapBridge(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
+    public static MapObstacleEntity makeMapBridge(float x, float y, LibGdxRenderingEntity libGdxRenderingEntity) {
         Image bridgeImage = ImageComponentFactory.makeImageComponent(x, y, 20, 5, new Texture("assets/MapObstacles/tree_trunk_square.png"));
         BoundingBoxComponent bridgeBoundingBoxComponent = BoundingBoxComponentFactory.makeBoundingBoxComponent(bridgeImage);
         libGdxRenderingEntity.getStage().addActor(bridgeImage);
-        return new MapObstacleComponent(bridgeBoundingBoxComponent, bridgeImage);
+        return new MapObstacleEntity(bridgeBoundingBoxComponent, bridgeImage);
     }
 }
